backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Flask API with PostgreSQL to serve time budgeting data"
requires-python = ">=3.9"
dependencies = [
    "Flask>=3.0.0",
    "Flask-SQLAlchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "Flask-Migrate>=4.0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]